split DocumentEditor/lists into shared/ui
