+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -30,7 +30,7 @@ import { clearFormatting, type Mark } from './utils'
 import { Toolbar } from './Toolbar'
 import { renderElement } from './render-element'
 import { withHeading } from './heading'
-import { nestList, unnestList, withList } from './lists'
+import { nestList, unnestList, withList } from './lists-shared'
 import { ComponentBlockContext, withComponentBlocks } from './component-blocks'
 import { getPlaceholderTextForPropPath } from './component-blocks/utils'
 import { withBlockquote } from './blockquote-shared'
